Make width element attribute work for the VideoPress iframe.
This commit aims to *not* apply width styling to the VideoPress iframe, and does so by making an exception with the negation css pseudo-class.

Every iframe that is a descendant of a .wpview-wrap class element gets 100% width style applied. This should not happen for the VideoPress iframe for two reasons:

1) The user defined w (width) shortcode attribute has to take effect.
2) The preloaded https://videopress.com/videopress-iframe.js script takes care of automatic resizing of the iframe by manipulating the width element attribute.